feat(order): add Order class
Orders are trays that indicate which food items to place and the score granted once the order is fulfilled. Completing orders is the game's primary method of accumulating score. Orders move along the conveyor belt. Missing an order once doesn't mean it's lost as it will return when the belt completes a full loop.